// src/shared/enums/exceptions.enum.ts
export enum Exception {
  FORBIDDEN = 'FORBIDDEN',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TRANSACTION_CATEGORY_ALREADY_EXISTS = 'TRANSACTION_CATEGORY_ALREADY_EXISTS',
  TRANSACTION_CATEGORY_NOT_FOUND = 'TRANSACTION_CATEGORY_NOT_FOUND',
  TRANSACTION_FOLDER_NOT_FOUND = 'TRANSACTION_FOLDER_NOT_FOUND',
  TRANSACTION_TYPE_NOT_FOUND = 'TRANSACTION_TYPE_NOT_FOUND',
  UNAUTHORIZED = 'UNAUTHORIZED',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
}